Re-organize api module with internal package
