+++ api/src/main/java/com/vimeo/networking2/GrantType.kt
@@ -1,4 +1,4 @@
-package com.vimeo.networking2.enums
+package com.vimeo.networking2
 
 /**
  * Grant types.
