+++ api/src/main/java/com/vimeo/networking2/config/RetrofitSetupModule.kt
@@ -3,7 +3,7 @@ package com.vimeo.networking2.config
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.adapters.Rfc3339DateJsonAdapter
 import com.vimeo.networking2.ApiConstants
-import com.vimeo.networking2.adapters.ErrorHandlingCallAdapterFactory
+import com.vimeo.networking2.internal.ErrorHandlingCallAdapterFactory
 import okhttp3.Interceptor
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
