+++ api/src/main/java/com/vimeo/networking2/internal/ErrorHandlingCallAdapterFactory.kt
@@ -1,4 +1,4 @@
-package com.vimeo.networking2.adapters
+package com.vimeo.networking2.internal
 
 import com.vimeo.networking2.ApiError
 import retrofit2.CallAdapter
