+++ auth/src/main/java/com/vimeo/networking2/Authenticator.kt
@@ -2,7 +2,6 @@ package com.vimeo.networking2
 
 import com.vimeo.networking2.config.ServerConfig
 import com.vimeo.networking2.requests.AuthCallback
-import com.vimeo.networking2.requests.VimeoRequest
 
 /**
  * API that allow you to make the following authentication requests:
@@ -18,6 +17,7 @@ import com.vimeo.networking2.requests.VimeoRequest
  *
  * Ex:
  *
+ * ```
  * val authenticator = Authenticator.create(serverConfig)
  * authenticator.clientCredentials(object: AuthCallback() {
  *
@@ -37,6 +37,7 @@ import com.vimeo.networking2.requests.VimeoRequest
  *
  *       }
  * })
+ * ```
  *
  */
 interface Authenticator {
