+++ example-vimeo-networking2-android/src/main/java/com/vimeo/vimeonetworking2/playground/MainActivity.kt
@@ -3,9 +3,9 @@ package com.vimeo.vimeonetworking2.playground
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import com.vimeo.moshiexampleandroid.R
+import com.vimeo.networking2.ApiResponse
 import com.vimeo.networking2.Authenticator
 import com.vimeo.networking2.config.ServerConfig
-import com.vimeo.networking2.requests.ApiResponse
 import com.vimeo.networking2.requests.AuthCallback
 import kotlinx.android.synthetic.main.activity_main.*
